using config for environment variables for development

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,4 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-
 import { UUID, randomUUID } from 'crypto';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDTO } from './dto/get-tasks-filter.dto';
@@ -11,7 +10,6 @@ import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
-
 
     constructor(
         @InjectRepository(TaskRepository)
@@ -25,7 +23,7 @@ export class TasksService {
 
     async getTaskById(id: number, user: User): Promise<Task> {
 
-        const found = await this.taskRepository.findOne({ where: { id , userId: user.id } });
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
         if (!found) {
             throw new NotFoundException(`Task with id: ${id} not found`);
         }
@@ -38,20 +36,18 @@ export class TasksService {
 
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const task = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const task = await this.taskRepository.delete({ id, userId: user.id });
         if (task.affected === 0) {
             throw new NotFoundException(`Task with id: ${id} not found`);
         };
+    };
 
-    }
-/*
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
 
     }
-*/
 }
